Add files via upload
This portfolio is a personal website created by Bhaibhav Raj, a MERN Full Stack Developer. It serves as a digital representation of Bhaibhav's skills, work, and professional identity. Here's a brief overview of its purpose and creation:

Purpose:

The portfolio's main purpose is to showcase Bhaibhav Raj's skills and expertise as a MERN Full Stack Developer.
It provides an introduction to his background, skills, and passion for web development.
The portfolio offers visitors the opportunity to view his past projects and get in touch for potential collaborations or hiring.
Importance:

A personal portfolio is crucial for professionals in the tech industry as it acts as a digital resume and showcases their capabilities.
It helps establish a strong online presence and builds credibility in the eyes of potential clients, employers, or collaborators.
A well-designed portfolio can set Bhaibhav apart from others in the field and leave a lasting impression on visitors.
Creation:

The portfolio is created using HTML for structuring the content and CSS for styling.
It employs modern design principles, animations, and a responsive layout to ensure a positive user experience across various devices.
External libraries like Boxicons are used for icons, enhancing the visual appeal of the site.
The portfolio is hosted online, likely on a web hosting service or platform for easy accessibility.